Use shared mixin to include help tab in all app types
