+++ app/static/src/app/components/fit/FitApp.vue
@@ -15,6 +15,9 @@
     </template>
     <template v-slot:right>
       <wodin-tabs id="right-tabs" :tabNames="rightTabNames" @tabSelected="rightTabSelected">
+        <template v-if="helpTabName" v-slot:[helpTabName]>
+          <help-tab></help-tab>
+        </template>
         <template v-slot:Run>
           <run-tab></run-tab>
         </template>
@@ -37,11 +40,13 @@ import WodinTabs from "../WodinTabs.vue";
 import CodeTab from "../code/CodeTab.vue";
 import DataTab from "../data/DataTab.vue";
 import FitTab from "./FitTab.vue";
+import HelpTab from "../help/HelpTab.vue";
 import RunTab from "../run/RunTab.vue";
 import OptionsTab from "../options/OptionsTab.vue";
 import SensitivityTab from "../sensitivity/SensitivityTab.vue";
 import { VisualisationTab } from "../../store/appState/state";
 import { AppStateMutation } from "../../store/appState/mutations";
+import includeHelpTab from "../mixins/includeHelpTab";
 
 export default defineComponent({
     name: "FitApp",
@@ -49,6 +54,7 @@ export default defineComponent({
         CodeTab,
         DataTab,
         FitTab,
+        HelpTab,
         RunTab,
         OptionsTab,
         SensitivityTab,
@@ -58,12 +64,15 @@ export default defineComponent({
     setup() {
         const store = useStore();
 
-        const rightTabNames = [VisualisationTab.Run, VisualisationTab.Fit, VisualisationTab.Sensitivity];
         const rightTabSelected = (tab: string) => {
             store.commit(AppStateMutation.SetOpenVisualisationTab, tab);
         };
+        const { helpTabName, rightTabNames } = includeHelpTab([
+            VisualisationTab.Run, VisualisationTab.Fit, VisualisationTab.Sensitivity
+        ]);
 
         return {
+            helpTabName,
             rightTabNames,
             rightTabSelected
         };
