+++ app/static/src/app/components/mixins/includeHelpTab.ts
@@ -1,0 +1,26 @@
+import {useStore} from "vuex";
+import {computed} from "vue";
+
+export default (nonHelpTabNames: string[]) => {
+    const store = useStore();
+
+    const helpTabName = computed(() => {
+        if (store.state.config?.help?.markdown?.length) {
+            return store.state.config.help.tabName || "Explanation"; // default if markdown but no tab name
+        }
+        // do not show tab if no markdown
+        return null;
+    });
+    const rightTabNames = computed(() => {
+        const result = [...nonHelpTabNames];
+        if (helpTabName.value) {
+            result.unshift(helpTabName.value);
+        }
+        return result;
+    });
+
+    return {
+        helpTabName,
+        rightTabNames
+    };
+};
