+++ config/help/day2.md
@@ -1,0 +1,10 @@
+## Example Help for WODIN app
+
+Here is an example formula which can be included in markdown help for an app:
+
+$$\frac{dS}{dt}=-\beta \frac{SI}{N}+\nu I$$
+
+And here is an example of an inline maths expression: $((\\beta I dt)/N)$
+
+Here is an example image, which should be placed in the `help/config` folder: ![example image](/example_img.png)
+
